Density of the multivariate normal and t distributions

Density of the multivariate normal and t distributions

Usage

dmvnorm_eclairs(x, mu, ecl, log = FALSE)

Arguments

x

numeric matrix with the data. The rows correspond to observations and the columns to variables.

mu

mean vector

ecl

covariance matrix as an eclairs object

nu

degrees of freedom, defaults to Inf. If finite, uses a multivariate t distribution

Value

A numerical vector with the density values calculated at each vector (row of the matrix x).

See also

Examples

library(Rfast)

X <- matrnorm(10000, 10)
mu <- colmeans(X)
Sigma <- cov(X)
ecl <- eclairs(X, lambda=1e-8)

# Multivariate normal
dmvnorm(X[1,], mu, Sigma, log=TRUE) 
#> [1] -11.20204
dmvnorm_eclairs( X[1,], mu, ecl, log=TRUE)
#> [1] -11.20204

# Multivariate Student t
dmvt(X[1,], mu, Sigma, nu=5, log=TRUE) 
#> [1] -10.95576
dmvt_eclairs( X[1,], mu, ecl, nu=5,  log=TRUE)
#> [1] -10.95576