Evaluate effective dependence

Evaluate effective dependence of the correlation matrix

effDependence(Sigma.eclairs)

Arguments

Sigma.eclairs
estimate of correlation matrix from eclairs storing \(U\), \(d_1^2\), \(\lambda\) and \(\nu\)

Details

Compute a metric of the average degree of dependence between all pairs of variables. Proposed by Pena and Rodriguez (2003), the 'effective dependence' is \(1-|C|^\frac{1}{p}\) where \(C\) is a correlation matrix between \(p\) variables. In the case where all pairwise correlations are \(\rho\), the effective dependence converges to \(\rho\) for sufficiently large \(p\). Like the squared correlation, the value is bounded between 0 and 1.

References

Pena D, Rodriguez J (2003). “Descriptive measures of multivariate scatter and linear dependence.” Journal of Multivariate Analysis, 85(2), 361--374.

Examples

library(Rfast)
set.seed(1)
n = 2000 # number of samples
p = 1000 # number of features
rho = .9

# create correlation matrix
Sigma = matrix(rho, p, p)
diag(Sigma) = 1

# draw data from correlation matrix Sigma
Y = rmvnorm(n, rep(0, p), sigma=Sigma*5.1)
rownames(Y) = paste0("sample_", 1:n)
colnames(Y) = paste0("gene_", 1:p)

# eclairs decomposition
Sigma.eclairs = eclairs(Y, compute="cor")

# effective variance
effVariance( Sigma.eclairs )

#> [1] 0.1200975

# effective dependence
effDependence( Sigma.eclairs )

#> [1] 0.8799025

# expected value with constant correlation
rho

#> [1] 0.9