Fit linear model on each feature after decorrelating

Fit linear model on each features after applying decorrelation projection to reponse and predictors.

lm_each_eclairs(
  formula,
  data,
  X,
  Sigma.eclairs,
  subset,
  weights,
  na.action,
  method = "qr",
  model = TRUE,
  x = FALSE,
  y = FALSE,
  qr = TRUE,
  singular.ok = TRUE,
  contrasts = NULL,
  offset,
  ...
)

Arguments

formula

an object of class 'formula' (or one that can be coerced to that class): a symbolic description of the model to be fitted.

data

a matrix or data.frame containing the variables in the model

X

matrix or data.frame where each column stores a predictor to be evaluated by the regression model one at a time. The \(i^{th}\) model includes X[,i] as a predictor.

Sigma.eclairs

estimate of covariance/correlation matrix from eclairs storing \(U\), \(d_1^2\), \(\lambda\) and \(\nu\)

subset

same as for lm

weights

same as for lm

na.action

same as for lm

method

same as for lm

model

same as for lm

x

same as for lm

y

same as for lm

qr

same as for lm

singular.ok

same as for lm

contrasts

same as for lm

offset

same as for lm

...

other arguments passed to lm()

Value

data.frame with columns beta, se, tsat, pvalue storing results for regression model fit for each feature

Examples

library(Rfast)
set.seed(1)
n = 800 # number of samples
p = 200 # number of features

# create correlation matrix
Sigma = autocorr.mat(p, .9)

# draw data from correlation matrix Sigma
Y = rmvnorm(n, rep(0, p), sigma=Sigma*5.1)

# eclairs decomposition
ecl = eclairs(Y)

# simulate covariates
data = data.frame(matrnorm(p,2))
colnames(data) = paste0('v', 1:2)

# simulate response
y = rnorm(p)

# Simulate 1000 features to test
X = matrnorm(p, 1000)
colnames(X) = paste0('set_', 1:ncol(X))

# Use linear model to test each features stored as columns in X
res = lm_each_eclairs(y ~ v1 + v2, data, X, ecl )

head(res)
#>              beta         se      tstat    pvalue
#> set_1  0.09346521 0.05884551  1.5883152 0.1138269
#> set_2 -0.01056301 0.06031143 -0.1751411 0.8611496
#> set_3 -0.08070951 0.06195747 -1.3026598 0.1942194
#> set_4 -0.07576136 0.06608938 -1.1463469 0.2530492
#> set_5  0.05135574 0.06294026  0.8159442 0.4155226
#> set_6 -0.03525925 0.07006731 -0.5032197 0.6153749

# Analysis after non-linear transform
#------------------------------------

# Apply function to transforme data
f = function(x) log(x^2 + 0.001)

# evaluate covariance of transformed data
ecl_transform = cov_transform(ecl, f, 100)

# Use linear model to test each features stored as columns in X
# in data transformed by f()
res2 = lm_each_eclairs( f(y) ~ v1 + v2, data, X, ecl_transform )

head(res)
#>              beta         se      tstat    pvalue
#> set_1  0.09346521 0.05884551  1.5883152 0.1138269
#> set_2 -0.01056301 0.06031143 -0.1751411 0.8611496
#> set_3 -0.08070951 0.06195747 -1.3026598 0.1942194
#> set_4 -0.07576136 0.06608938 -1.1463469 0.2530492
#> set_5  0.05135574 0.06294026  0.8159442 0.4155226
#> set_6 -0.03525925 0.07006731 -0.5032197 0.6153749