Compute SVD with selected algorithm

Compute SVD using exact, approximate or randomized algorithms

Usage

run_svd(X, k = min(dim(X)), method = c("svd", "irlba", "pcaone"), ...)

Arguments

X

data matrix

k

number of left and right singular values to compute

method

SVD algorithm string "svd", "irlba", or "pcaone"

...

other arguments passed to SVD functions

Details

Compute SVD using select algorithm. Let data matrix X have n rows and p columns. At most min(n,p) singular vectors can be computed.

"svd": exact algorithm using standard svd() function. Fastest choice if computing greater than about min(n,p)/3 singular vectors.

"irlba": approximate algorithm using implicitly restarted Lanczos bidiagonalization in irlba::irlba(). Much faster than "svd" for computing less than about min(n,p)/3 singular vectors, while matching svd() to moderate precision

"pcaone": randomized SVD algorithm implemented in pcaone::pcaone(). Fastest method for large datasets, when computing about min(n,p)/3 singular vectors. Singular vectors and values are approximate and can differ from svd() by a few percent.

Examples

hilbert <- function(n) { i <- 1:n; 1 / outer(i - 1, i, `+`) }
X <- hilbert(9)[, 1:6]
dcmp1 <- run_svd(X, method = "svd")
dcmp2 <- run_svd(X, method = "irlba")
dcmp3 <- run_svd(X, method = "pcaone")