Get full covariance/correlation matrix from eclairs

Get full covariance/correlation matrix from eclairs decomposition

getCov(Sigma.eclairs, lambda, ...)

getCor(Sigma.eclairs, lambda, ...)

# S4 method for eclairs
getCov(Sigma.eclairs, lambda, ...)

# S4 method for eclairs
getCor(Sigma.eclairs, lambda, ...)

Arguments

Sigma.eclairs
eclairs decomposition

lambda
shrinkage parameter for the convex combination.

...
other arguments

Value

p x p covariance/correlation matrix

Details

The full matrix is computationally expensive to compute and uses a lot of memory for large p. So it is better to use decorrelate or mult_eclairs to perform projections in \(O(np)\) time.

Examples

library(Rfast)
set.seed(1)
n = 800 # number of samples
p = 200 # number of features

# create correlation matrix
Sigma = autocorr.mat(p, .9)

# draw data from correlation matrix Sigma
Y = rmvnorm(n, rep(0, p), sigma=Sigma*5.1)
rownames(Y) = paste0("sample_", 1:n)
colnames(Y) = paste0("gene_", 1:p)

# eclairs decomposition
Sigma.eclairs = eclairs(Y)

# extract covariance implied by eclairs decomposition
getCov(Sigma.eclairs)[1:3, 1:3]

#>          [,1]     [,2]     [,3]
#> [1,] 5.177183 4.496313 4.004525
#> [2,] 4.496313 5.132412 4.506282
#> [3,] 4.004525 4.506282 5.148518