Recompute eclairs after dropping features

Recompute eclairs after dropping features

reform_decomp(Sigma.eclairs, k = Sigma.eclairs$k, drop = NULL)

Arguments

Sigma.eclairs
covariance/correlation matrix as an eclairs object

k
the rank of the low rank component

drop
array of variable names to drop.

Value

eclairs decomposition for a subset of features

Details

Reform the dataset from the eclairs decomposition, drop features, then recompute the eclairs decomposition. If the original SVD/eigen was truncated, then the reconstruction of the original data will be approximate. Note that the target shrinkage matrix is the same as in Sigma.eclairs, so \(\nu\) is not recomputed from the retained features.

Examples

library(Rfast)
set.seed(1)
n = 800 # number of samples
p = 200 # number of features

# create correlation matrix
Sigma = autocorr.mat(p, .9)

# draw data from correlation matrix Sigma
Y = rmvnorm(n, rep(0, p), sigma=Sigma*5.1)
rownames(Y) = paste0("sample_", 1:n)
colnames(Y) = paste0("gene_", 1:p)

# Correlation
#------------

# eclairs decomposition
Sigma.eclairs = eclairs(Y, compute="correlation")

# features to drop
drop = paste0("gene_",1:100)

# Compute SVD on subset of eclairs decomposition
ecl1 = reform_decomp( Sigma.eclairs, drop=drop)

ecl1

#>        Estimate covariance/correlation with low rank and shrinkage
#> 
#>   Original data dims: 800 x 100 
#>   Low rank component: 100 
#>   lambda:             0.0268 
#>   nu:                 1 
#>   Avg corr (MLE):     0.989 
#>   Avg corr (EB):      0.962 
#>   logML:              211207